Modification créa campagne et insert admin

+++ FabApp/database/migrations/2023_04_03_000001_create_usager_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('nom');
             $table->string('adresseCourriel');
             $table->text('motDePasse');
-            $table->float('role')->default(2);
+            $table->int('role')->default(2);
         });
     }
